Refactored input
