+++ packages/fakeit-core/package.json
@@ -23,14 +23,7 @@
     "debug": "^3.1.0",
     "faker": "^4.1.0",
     "relieve": "^2.2.1",
-    "yamljs": "^0.3.0",
-    "@fakeit/input": "^1.0.0"
-  },
-  "optionalDependencies": {
-    "@fakeit/input": "^1.0.0"
-  },
-  "devDependencies": {
-    "@fakeit/input": "^1.0.0"
+    "@fakeit/file-loader": "^1.0.0"
   },
   "plugins": [
     "@fakeit/format-csv",
