+++ packages/fakeit-core/src/input.js
