--- packages/fakeit-input/.npmignore
@@ -1,3 +1,0 @@
-src
-test
-*.log
