--- packages/fakeit-input/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@fakeit/input",
-  "description": "FakeIt Input",
-  "author": "Aaron Benton <bentonam@gmail.com> (https://github.com/bentonam)",
-  "contributors": [
-    "Tyler Benton <tjbenton21@gmail.com> (https://github.com/tjbenton)"
-  ],
-  "version": "1.0.0",
-  "main": "./dist/index.js",
-  "repository":
-    "https://github.com/bentonam/fakeit/tree/master/packages/fakeit-input",
-  "bugs": {
-    "url": "https://github.com/bentonam/fakeit/issues"
-  },
-  "keywords": [],
-  "license": "MIT",
-  "homepage": "https://github.com/bentonam/fakeit#readme",
-  "dependencies": {
-    "@fakeit/file-loader": "^1.0.0"
-  },
-  "peerDependencies": {
-    "debug": "^3.1.0"
-  },
-  "devDependencies": {
-    "debug": "^3.1.0"
-  }
-}
